+++ sequana/phred.py
@@ -1,26 +1,47 @@
-"""
+"""Manipulate phred quality of reads
 
-phred scales indicates the range of of characters.
-characters goes from ! to ~ that is tfrom 33 to 126in ascii tqble. Characters before ! caused trouble (e.g. white spaces). This scale is the Sanger scale. 2 other scales could be used ranging from 59 to 126 (illumina 1) and from 64 to 126 (illumina 1.3+).
+FastQ quality are stored as characters. The phred scales indicates the range of characters.
 
-
-http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2847217/
-
-Here qre the offset to use::
-
-Sanger : 33       range 0 to 93
-Solexa: 64        range -5 to 62
-illumina1.3+: 64  range 0 to 62
+In general, characters goes from ! to ~ that is from 33 to 126 in an ascii
+table. This convention starts at 33 because characters before ! may cause trouble 
+(e.g. white spaces). This scale is the Sanger scale. There are 2 other scales
+that could be used ranging from 59 to 126 (illumina 1) and from 64 to 126 (illumina 1.3+).
 
 
-::
 
-    p = linspace(0,1,1000)
-    plot(p, [phred.phred_quality_score_sanger(this)
-        for this in p], color='red' )
-    plot(p, [phred.phred_score_solexa(this)
-        for this in p], color='black' )
+So, here are the offset to use:
 
+============== ============ ===============
+ Name           offset       Numeric range
+============== ============ ===============
+ Sanger         33            0 to 93
+ Solexa         64            -5 to 62
+ illumina1.3+   64            0 to 62
+============== ============ ===============
+
+:reference: http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2847217/
+
+
+Even though dedicated tools would have better performances, we provide a set
+of convenient functions. An example is provided here below to plot the quality
+corresponding to a character string extracted from a FastQ read.
+
+
+In this example, we use :class:`Quality` class where the default offset is 33
+(Sanger). We compare the quality for another offset 
+
+.. plot::
+    :include-source:
+
+    from sequana import phred
+
+    from sequana.phred import Quality
+    q = Quality('BCCFFFFFHHHHHIIJJJJJJIIJJJJJJJJFH')
+    q.plot()
+    q.offset = 64
+    q.plot()
+    from pylab import legend
+    legend(loc="best")
 
 
 """
@@ -58,6 +79,7 @@ q low proba of errors (0.01) gives Q = 20
     return Qs
 
 def quality_to_proba_sanger(quality):
+    """Quality to probability (Sanger)"""
     return 10**(quality/-10.)
 
 
@@ -67,18 +89,16 @@ def proba_to_quality_solexa(pe):
     """
     if pe > 1:
         pe = 1
+        return -5
+
     if pe <1e-90:
         pe = 1e-90
     Qs = -10 * log10(pe/(1-pe))
     if Qs > 62:
         Qs = 62
-    if Q < -5:
-        Q = -5
+    if Qs < -5:
+        Qs = -5
     return Qs
-
-
-def quality_to_proba_sanger(quality):
-    return 10**(quality/-10.)
 
 
 def quality_solexa_to_quality_sanger(qual):
@@ -86,11 +106,17 @@ def quality_solexa_to_quality_sanger(qual):
 
 
 def quality_sanger_to_quality_solexa(qual):
+    """
+
+
+    """
     return 10 * log10(10**(qual/10.) - 1 )
 
 
 def ascii_to_quality(character, phred=33):
-    """
+    """ASCII to Quality conversion
+
+    :param int phred: offset (defaults to 33)
 
     ::
 
@@ -103,29 +129,31 @@ def ascii_to_quality(character, phred=33):
 
 
 def quality_to_ascii(quality, phred=33):
-    """
+    """Quality to ASCII conversion 
+
+    :param int phred: offset (defaults to 33)
+
     ::
 
          >>> quality_to_ascii(65)
          b
 
     """
-    return chr(quality+33)
+    return chr(quality + phred)
 
 
 class Quality(object):
-    """
-
+    """Phred quality
 
     .. plot::
+        :include-source:
 
-        >>> from sequana.phread import Quality
+        >>> from sequana.phred import Quality
         >>> q = Quality('BCCFFFFFHHHHHIIJJJJJJIIJJJJJJJJFH')
         >>> q.plot()
-        >>> q.mean_quality()
-        35
-        
 
+    You can access to the quality as a list using the :attr:`quality` attribute
+    and the mean quality from the :attr:`mean_quality` attribute.
 
     """
     def __init__(self, seq, offset=33):
@@ -135,30 +163,38 @@ class Quality(object):
     def _get_quality(self):
         # bytearray conversion is required since ord() function
         # does not handle bytes from py3
-        return [x - self.offset for x in bytearray(self.seq)]
-    quality = property(_get_quality)
+        return [x - self.offset for x in bytearray(self.seq, 'utf-8')]
+    quality = property(_get_quality, doc="phred string into quality list")
 
     def _get_mean_quality(self):
         return np.mean(self.quality)
-    mean_quality = property(_get_mean_quality)
+    mean_quality = property(_get_mean_quality, doc="return mean quality")
 
-    def plot(self):
-        pylab.plot(self.quality)
-        pylab.xlabel('base position')
-        pylab.ylabel('Quality per base')
-        pylab.grid()
+    def plot(self, fontsize=16):
+        """plot quality versus base position"""
+        pylab.plot(self.quality, label="offset: %s" % self.offset)
+        pylab.xlabel('base position', fontsize=fontsize)
+        pylab.ylabel('Quality per base', fontsize=fontsize)
+        pylab.grid(True)
+        # ylim set autoscale to off so if we want to call this function  several
+        # times, we must reset autoscale to on before calling ylim
+        pylab.autoscale()
+        limits = pylab.ylim()
+        pylab.ylim(max(0,limits[0]-1), limits[1]+1)
 
 
 
 # this should qlso be correct for Illumina 1.8+
 class QualitySanger(Quality):
-    def __init__(self, seq, offset=33):
-        super(QualitySanger, self).__init__(seq, offset)
+    """Specialised :class:`Quality` class for Sanger case"""
+    def __init__(self, seq):
+        super(QualitySanger, self).__init__(seq, offset=33)
 
 
 class QualitySolexa(Quality):
-    def __init__(self, seq, offset=64):
-        super(QualitySanger, self).__init__(seq, offset)
+    """Specialised :class:`Quality` class for Solexa case"""
+    def __init__(self, seq):
+        super(QualitySolexa, self).__init__(seq, offset=64)