+++ sequana/resources/data/__init__.py
@@ -1,0 +1,19 @@
+"""
+
+Some useful data sets to be used in the analysis
+
+
+
+The command :func:`sequana.sequana_data` may be used to retrieved data from 
+this package. For example, a small but standard reference (fix) is used in 
+some NGS experiments. The file is small enough that it is provided within
+sequana and its filename (full path) can be retrieved as follows::
+
+    from sequana import sequana_data
+    fullpath = sequana_data("Coliphage_phiX174.fa", "data")
+
+Other files stored in this directory will be documented here.
+
+
+
+"""