+++ doc/source/Changelog.rst
@@ -4,6 +4,19 @@ Changelog
 .. contents::
 
 
+0.1.0 April 2006
+------------------
+
+First functional release.
+
+* NEWS
+  - rules in sequana/rules and pipelines in sequana/pipelines 
+  - standalone tool called sequana to download a Snakefile + config in a local
+directory for further analysis
+  - modules for post-analysis: bamtools, vcf_filter, fastq, ....
+  - Fully tested (90% coverage) and documented
+  - Set of reports
+
 0.0.4 March 2016
 -------------------